nutriblend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
black==24.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-cors-headers==4.3.1
django-import-export==3.3.8
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
MarkupPy==1.14
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
sqlparse==0.5.0
tablib==3.5.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
